+++ src/components/UserInfo/UserInfo.tsx
@@ -1,19 +1,10 @@
 import { FC } from 'react';
+import type { User } from '../../types';
 
-type User = {
-  id: number;
-  name: string;
-  username: string;
-  email: string;
-};
 type Props = {
   user: User;
 };
 export const UserInfo: FC<Props> = ({ user }) => {
-  if (!user) {
-    return null;
-  }
-
   return (
     <a className="UserInfo" href={`mailto:${user.email}`}>
       {user.name}
